v0.3.1143 follow-on — SCALE-SEAM ㊻ ㊼ ㊽, verification / rent-roll / permits
Move eight methods out of client.ts into existing mixins (no extra withX):
field verification onto model.ts, in-place rent onto proforma.ts, open-data
permits onto entitlements.ts. Pin 2,127 → 2,065. MARKS widened to ㊾.
