+++ apps/web/src/api/entitlements.ts
@@ -1,5 +1,5 @@
 import { HttpCore } from "./httpCore";
-import type { EntitlementConditions, ReviewCycles } from "./types";
+import type { EntitlementConditions, OpendataPermit, ReviewCycles } from "./types";
 
 /**
  * Entitlements — the `/projects/{pid}/entitlements/…` route group.
@@ -13,6 +13,10 @@ import type { EntitlementConditions, ReviewCycles } from "./types";
  * it" — from opposite ends: `entitlementReviewCycles` measures the time an application has already
  * spent and whose court held it; `entitlementConditions` tracks what an approval obliges us to do
  * afterwards. Neither ever reports a condition satisfied or a round scored on incomplete data.
+ *
+ * SCALE-SEAM ㊽ adds open-data municipal filings — *which permits exist near this site?*
+ * Cities catalog, query, import into the GC permit module. `permitReadiness` stayed — that
+ * is a model-submission checklist, not a city feed.
  */
 type Ctor<T> = new (...args: any[]) => T;
 
@@ -48,6 +52,25 @@ export function withEntitlements<TBase extends Ctor<HttpCore>>(Base: TBase) {
       refused: { ref: string | null; workflow_state: string }[];
       total_exceeds: number; total_not_checkable: number; note: string;
     }>(`/projects/${pid}/entitlements/condition-checks`);
+  }
+
+  /** permitCities — open-data permit sources this deployment can query. */
+  permitCities() {
+    return this.json<{ cities: { id: string; label: string; region: string; authority: string; geo: boolean }[] }>(
+      "/opendata/permit-cities");
+  }
+  /** opendataPermits — a city's filings near a point / by text. */
+  opendataPermits(pid: string, opts: { city: string; lat?: number; lon?: number; radius?: number; address?: string; q?: string; limit?: number }) {
+    const qs = new URLSearchParams({ city: opts.city });
+    for (const k of ["lat", "lon", "radius", "address", "q", "limit"] as const)
+      if (opts[k] !== undefined && opts[k] !== "") qs.set(k, String(opts[k]));
+    return this.json<{ city: string; count: number; permits: OpendataPermit[] }>(
+      `/projects/${pid}/opendata/permits?${qs}`);
+  }
+  /** importOpendataPermits — import a city's filings into the GC permit module (source-tagged, deduped). */
+  importOpendataPermits(pid: string, body: { city: string; lat?: number; lon?: number; radius?: number; address?: string; q?: string; max?: number }) {
+    return this.json<{ imported: number; skipped: number; found: number; refs: string[] }>(
+      `/projects/${pid}/opendata/permits/import`, { method: "POST", body: JSON.stringify(body) });
   }
   };
 }
