+++ apps/web/src/api/proforma.ts
@@ -7,6 +7,10 @@
  *  Reports (`reports` / `reportUrl`) sat immediately above the cluster and did **not** come —
  *  those are the document catalog, not valuation. Composed through the existing `withProforma`
  *  wrapper — no extra `withX()` on `ApiClient`.
+ *
+ *  SCALE-SEAM ㊼ adds in-place operations — *what is this asset earning today?* Rent roll and
+ *  lease-management depth. They sat above the investor stack in `client.ts` and did **not** go
+ *  with ㉙ (capital is ownership, this is occupancy). `askProject` stayed.
  */
 import { HttpCore } from "./httpCore";
 import type { Appraisal, FinancialStatements, MonteCarloResult, ProformaForecast, ProformaResult } from "./types";
@@ -224,6 +228,31 @@ export function withProforma<TBase extends Ctor<HttpCore>>(Base: TBase) {
     return this.json<{ target: string; remote_id: string | null; url: string | null;
       fields_pushed: number; status: string }>(
       `/projects/${pid}/listings/${lid}/syndicate`, { method: "POST" });
+  }
+
+  /** rentRoll — occupancy, WALT, expiration schedule, in-place income. */
+  rentRoll(pid: string) {
+    return this.json<{ occupancy_pct: number; lease_count: number; base_rent_annual: number;
+      in_place_gross_income: number; walt_years: number; expirations_by_year: Record<string, unknown>;
+      rows: Record<string, unknown>[] }>(`/projects/${pid}/rent-roll`);
+  }
+  /** leaseManagement — renewal pipeline, rent-escalation schedule, CAM/recovery reconciliation. */
+  leaseManagement(pid: string, years?: number, recoverableOpex?: number) {
+    const q = new URLSearchParams();
+    if (years != null) q.set("years", String(years));
+    if (recoverableOpex != null) q.set("recoverable_opex", String(recoverableOpex));
+    const qs = q.toString() ? `?${q}` : "";
+    return this.json<{
+      lease_count: number;
+      renewals: { holdover_count: number; expired_count: number; options_outstanding: number;
+        at_risk_rent: number; expiring: Record<string, { count: number; rent: number }>;
+        rows: Record<string, unknown>[] };
+      escalations: { years: number; portfolio_by_year: number[]; current_base_rent: number;
+        projected_base_rent: number; rows: Record<string, unknown>[] };
+      cam: { recoverable_income: number; recoverable_sf: number; by_lease_type: Record<string, number>;
+        recovery_ratio?: number | null; over_recovery?: number; under_recovery?: number;
+        rows: Record<string, unknown>[] };
+    }>(`/projects/${pid}/leases/management${qs}`);
   }
   };
 }
