+++ apps/web/src/api/surface.test.ts
@@ -162,6 +162,9 @@ describe("the API client's public surface", () => {
       "viewTemplates", "saveViewTemplates", "resolveViewTemplate",
       "selectionsSummary", "pushSelectionChangeEvents",
       "verifiedProgress", "progressRollup", "progressCaptureDiff",
+      "verificationCoverage", "setVerification", "verificationDeviations",
+      "rentRoll", "leaseManagement",
+      "permitCities", "opendataPermits", "importOpendataPermits",
       "clashMetrics", "clashClearanceRules", "clashMatrix",
       "chartOfAccounts", "namingConventions", "elementsByDiscipline", "pipelineFunnel",
       "twoSidedBudget", "mepPressureLoss", "mepTrayFill", "mepThermalLoads",
