+++ docs/roadmap.md
@@ -1188,7 +1188,7 @@ two rows share a path, so two agents in different rows cannot collide.
 | **F · Docs & demo** | `README.md`, `docs/`, `apps/web/src/demo/` | keep the shipped surface honest (below) — no coded items. **`demoData.test.ts` now gates the shell's startup endpoints**; re-run `build_demo_data.py` and that test after adding one |
 | **G · API surface** | `services/api/src/aec_api/routers/`, `main.py` | no standalone items: **every lane routes its own work**, which is why this is a lane rather than a shared file |
 | **H · Registers** | `services/api/modules/*/module.json` | — |
-| **I · API client** | `apps/web/src/api/` | SCALE-SEAM ㊻ *(the only open slice; ②–㊺ have shipped. This cell named ⑬–⑳ until 2026-08-24 — eight slices whose extractions had already landed — because the item regex could not see `㉒` at all, so nothing required this row to be right)* |
+| **I · API client** | `apps/web/src/api/` | SCALE-SEAM ㊾ *(the only open slice; ②–㊽ have shipped. This cell named ⑬–⑳ until 2026-08-24 — eight slices whose extractions had already landed — because the item regex could not see `㉒` at all, so nothing required this row to be right)* |
 | **J · Build & tooling** | `apps/web/scripts/`, `apps/web/vite.config.ts`, `apps/web/src/style.css`, `apps/web/src/tooling/`, `services/api/test_file_sizes.py`, `services/api/run_tests.py` | R39-TSC-CACHE *(local typecheck once diverged from CI; cause unknown, prior explanation retracted — an OBSERVATION, not a defect with a known fix. Read the entry before "fixing" it: the proposed fix is named there and rejected)* |
 
 **Parked — not available to pick up.** These are decisions or multi-release commitments, listed so
@@ -3061,7 +3061,25 @@ verbs, with a command bar as the escape hatch to everything); and **role-shaped 
 
 ## 🧱 Decomposition & reliability carry-overs (interleave one per few releases)
 
-- ◧ ⭐ **SCALE-SEAM ㊻ — `client.ts` is no longer a god-file, but the split is not finished.** *(②–㊺ have shipped, **㊸–㊺ in the v0.3.1143 follow-on** — no version bump, tag lag already at the `test_release_current` bound)*
+- ◧ ⭐ **SCALE-SEAM ㊾ — `client.ts` is no longer a god-file, but the split is not finished.** *(②–㊽ have shipped, **㊻–㊽ in the v0.3.1143 follow-on** — no version bump, tag lag already at the `test_release_current` bound)*
+  **㊽ took open-data municipal filings out** (three methods) into the existing
+  `apps/web/src/api/entitlements.ts` mixin — **no extra wrapper on `ApiClient`**. Grouped by what
+  they ANSWER: *which permits exist near this site?* Cities catalog, query, import.
+  **`permitReadiness` did NOT come** — that is a model-submission checklist.
+
+  **㊼ took in-place operations out** (two methods) into the existing
+  `apps/web/src/api/proforma.ts` mixin — **no extra wrapper on `ApiClient`**. Grouped by what
+  they ANSWER: *what is this asset earning today?* Rent roll and lease-management depth.
+  They sat above the investor stack and did **not** go with ㉙. **`askProject` did NOT come.**
+
+  **㊻ took field-install verification out** (three methods) into the existing
+  `apps/web/src/api/model.ts` mixin — **no extra wrapper on `ApiClient`**. Grouped by what they
+  ANSWER: *is this element installed as designed?* Coverage, status write, deviation log.
+  They were **not** contiguous. **The photo upload did NOT come** (PHOTO-PIN is parked).
+  **`askModel` did NOT come.**
+  **`MARKS` was widened to ㊾ on the way out.** **The per-file pin moved 2,127 → 2,065**
+  across ㊻+㊼+㊽.
+
   **㊺ took the progress / as-built stack out** (three methods) into the existing
   `apps/web/src/api/schedule.ts` mixin — **no extra wrapper on `ApiClient`**. Grouped by what they
   ANSWER: *is claimed progress the installed one?* Verified vs claimed, GUID rollup, capture
